Initial working commit
This is the initial working commit of waf support for building
Arduino software. This has only been tested on Linux so far. If
it doesn't work on Windows (I plan to test sometime), I don't
think it'd be too hard to fix.

To customize the files for a different project, you should just
need to change config.ini. wscript and the new waf-tools should
be setup to read everything needed from config.ini.